Implemented Change Download and Application with Save

+++ Codemash/Phone/Codemash.Phone.Data/Repository/IRepository.cs
@@ -20,6 +20,6 @@ namespace Codemash.Phone.Data.Repository
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        T Get(int id);
+        T Get(long id);
     }
 }
